Change `MarkandSweepGC` CTOR API from `new_spawn_test_dummies()` to `new_with_test_dummy_roots()`
The old name to the CTOR method in `MarkandSweepGC` did not reflect that
the caller of the CTOR was creating a new `MarkandSweepGC` with dummy
roots. Rather, it was more vague and suggested general `DummyObjects`.
This is an important distinction since the roots, while being
`DummyObject`s, have the special invarient that they are created at
`MarkandSweepGC`'s initialization and act as the heads to appended
`DummyObject`s.